Update EPCglobal-epcis-2_0.xsd
This proposed change is for relaxing the XSD requirement that the Standard Business Document Header (SBDH) SHALL be present IF the EPCISHeader is included for other purposes such as declaring trade item master data attributes. To close out Issue 290.

Forgive me if I am submitting this in an unsuitable way.